SAT assignment and assignment checking
Added assignment to represent a set of the booleans to check satisfication.
Pumped up problem to include several new things:
 Added a map from node to vertex_descriptor to access graph using node.
 Added connected_components functionality to deal with multiple connected graphs.
 Added assignment satisfication checks: set of clauses satisfied, number satisfied, and if all are satisfied.
Added satisfiability_visitor, a visitor that loops through graph and collects clause satisfication.
Can generate solvable 3sat problem in problem_factory.
Added tests for satisfiability in main and testing.
Removed boost aliases in bgl_inc.
Moved documentation into folder "doc" and added build documentation for boost.
Changed project to require boost to be in the same folder as coopnet.

+++ src/sat/component/node.cpp
@@ -22,7 +22,7 @@ bool sat::operator==(const node& node_1, const node& node_2) {
 
 node_list::node_list() {}
 
-node_list::node_list(int num_nodes) {
+node_list::node_list(size_t num_nodes) {
 	for (auto i = 0; i < num_nodes; ++i) nodes.emplace_back(i);
 }
