+++ src/sat/component/node.h
@@ -35,7 +35,7 @@ namespace sat {
 		
 		node_list();
 
-		node_list(int num_nodes);
+		node_list(size_t num_nodes);
 
 		template<typename iter>
 		explicit node_list(iter begin, iter end)
